Tolkien-related bibliographies: Nouvelle ecole...
"Nouvelle École" is a far-right French intellectual journal,
affiliated with GRECE, an extreme far-right think tank of the
"Nouvelle Droite" movement.
Known for its nationalist and identitarian positions, it attracts
contributions from various scholars, which can lend a veneer of
intellectual respectability to its ideological positions.
This bibliography is provided for information purposes only.
Some well-known Tolkien scholars have published in this magazine.
The compiler does not endorse the views of the magazine.
Readers should be aware of its ideological leanings.