fix(bibliographies): Remove / comment out unwanted content from Mallorn
We had done a pass in the past, but there was still a lot of these left:
Journal editorials, lettres from readers, fan poems, fan fiction book or
article not related to Tolkien studies, etc.
The aim is to only keep articles and essays relevant for Tolkien studies
in a broad sense, and such entries certainly not belong to that category.